requirements.txt
asgiref==3.5.2
certifi==2022.6.15
coverage==6.4.4
Django==4.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
